Py 2 to 3 conversion

+++ capture_gui/lib.py
@@ -393,4 +393,4 @@ def get_maya_main_window():
     """
     ptr = omui.MQtUtil.mainWindow()
     if ptr is not None:
-        return shiboken.wrapInstance(long(ptr), QtWidgets.QWidget)
+        return shiboken.wrapInstance(int(ptr), QtWidgets.QWidget)
